Serve assets thru webpack-dev-middleware
Configured to serve the assets and render the backend through
express. We're not using a view library, instead just sending a string
at the root. This should be enough for now, but we can configure
SSR with React if we want. Given that this is a game, I'm not sure
that's valuable here.

HMR is enabled, but App.js will need to be wrapped with
`react-hot-loader` to function correctly for React. For now, it does
a page refresh but does inject the styles at runtime.

The express generator is used as the base for this server. We also
kill off the polyfills, as we're only targeting new browsers. We will also
update the babel configuration to do so as well.