Extract bootstrapping to run function & delta
This loosely follows the structure of `brookjs`, except we've replaced
Redux w/ a `#scan`. We also implement some queuing logic to ensure
the actions are fired in the correct order, a problem we have currently
in `observeDelta`.

This is just setting up some scaffolding so we can maintain the game
state on the server and seggregate the managing socket connections.